requirements.txt
django==2.2.1
channels==2.2.0
psycopg2-binary==2.8.2
djangorestframework==3.9.4
django_extensions=2.1.6
